New branch FedoraEZService-mods created for modified stylesheets which can output MODS as well as FOXML.
The stylesheets have changed quite a bit. Instead of one single template containing the whole of the output document structure and lots of for-each loops, I've used multiple templates, for the different
output formats, datastreams (shared by the different output formats) and for repeatable elements.

As most of the information (apart from the datastreams) in both FOXML and MODS is in attributes, I've used attribute-sets for the static attributes, and attribute value templates for the dynamic ones.

A stylesheet containing the code common to both the SDef and SDep stylesheets (ezcommon.xsl) is used.

By default, the stylesheets will still output FOXML. A param 'format', with value 'mods', will cause MODS to be output.